+++ cdflib/xarray/cdf_to_xarray.py
@@ -41,10 +41,6 @@ def _convert_cdf_time_types(
 
     data = np.atleast_1d(np.squeeze(data))
 
-    if to_datetime and to_unixtime:
-        logger.info("Cannot convert to both unixtime and datetime.  Continuing with conversion to unixtime.")
-        to_datetime = False
-
     # Convert all data in the "data" variable to unixtime or datetime if needed
     data_type = properties.Data_Type_Description
     if len(data) == 0 or data_type not in ("CDF_EPOCH", "CDF_EPOCH16", "CDF_TIME_TT2000"):
@@ -66,21 +62,39 @@ def _convert_cdf_time_types(
 
     # Convert all the attributes in the "atts" dictionary to unixtime or datetime if needed
     new_atts = {}
+    time_converted_attrs = []
     for att in atts:
-        data_type = atts[att].Data_Type
+        attr_data_type = atts[att].Data_Type
         data = atts[att].Data
-        if data_type not in ("CDF_EPOCH", "CDF_EPOCH16", "CDF_TIME_TT2000"):
+        if attr_data_type not in ("CDF_EPOCH", "CDF_EPOCH16", "CDF_TIME_TT2000"):
             new_atts[att] = data
         else:
             if to_datetime:
+                time_converted_attrs.append(att)
                 new_atts[att] = cdfepoch.to_datetime(data)
             elif to_unixtime:
+                time_converted_attrs.append(att)
                 new_atts[att] = cdfepoch.unixtime(data)
             else:
-                if data_type == "CDF_EPOCH16":
+                if attr_data_type == "CDF_EPOCH16":
                     new_atts[att] = cdfepoch.compute(cdfepoch.breakdown(data)[0:7])
                 else:
                     new_atts[att] = data
+
+    # This is a bit of a hack, these data types are ambiguous
+    # Lets add an attribute so at least we retain some information about what these numbers represent
+    if data_type in ("CDF_EPOCH", "CDF_REAL4", "CDF_REAL8"):
+        new_atts["CDF_DATA_TYPE"] = data_type
+
+    # Another hack for now, if we convert the data to datetime or unixtime we want to have a memory of what it was before
+    if data_type in ("CDF_EPOCH", "CDF_EPOCH16") and (to_datetime or to_unixtime):
+        new_atts["CDF_DATA_TYPE"] = data_type
+    elif data_type == "CDF_TIME_TT2000" and to_unixtime:
+        new_atts["CDF_DATA_TYPE"] = data_type
+
+    # Yet another hack, if we convert the attributes too we need to somehow keep track of what was converted
+    if time_converted_attrs:
+        new_atts["TIME_ATTRS"] = time_converted_attrs
 
     return new_data, new_atts
 
@@ -289,10 +303,16 @@ def _convert_fillvals_to_nan(var_data: npt.NDArray, var_atts: Dict[str, Any], va
         ):
             if new_data.size > 1:
                 if new_data[new_data == var_atts["FILLVAL"]].size != 0:
-                    new_data[new_data == var_atts["FILLVAL"]] = np.nan
+                    if new_data.dtype.type == np.datetime64:
+                        new_data[new_data == var_atts["FILLVAL"]] = np.datetime64("nat")
+                    else:
+                        new_data[new_data == var_atts["FILLVAL"]] = np.nan
             else:
                 if new_data == var_atts["FILLVAL"]:
-                    new_data = np.array(np.nan)
+                    if new_data.dtype.type == np.datetime64:
+                        new_data[new_data == var_atts["FILLVAL"]] = np.array(np.datetime64("nat"))
+                    else:
+                        new_data[new_data == var_atts["FILLVAL"]] = np.array(np.nan)
     return new_data
 
 
@@ -666,7 +686,7 @@ def _verify_dimension_sizes(created_data_vars: Dict[str, xr.Variable], created_c
                     )
 
 
-def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = False, fillval_to_nan: bool = False) -> xr.Dataset:
+def cdf_to_xarray(filename: str, to_datetime: bool = True, to_unixtime: bool = False, fillval_to_nan: bool = False) -> xr.Dataset:
     """
     This function converts CDF files into XArray Dataset Objects.
 
@@ -681,7 +701,7 @@ def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = 
 
     Example MMS:
         >>> # Import necessary libraries
-        >>> import cdflib
+        >>> import cdflib.xarray
         >>> import xarray as xr
         >>> import os
         >>> import urllib.request
@@ -693,7 +713,7 @@ def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = 
         >>>     urllib.request.urlretrieve(url, fname)
 
         >>> # Load in and display the CDF file
-        >>> mms_data = cdflib.cdf_to_xarray("mms2_fgm_srvy_l2_20160809_v4.47.0.cdf", to_unixtime=True, fillval_to_nan=True)
+        >>> mms_data = cdflib.xarray.cdf_to_xarray("mms2_fgm_srvy_l2_20160809_v4.47.0.cdf", to_unixtime=True, fillval_to_nan=True)
 
         >>> # Show off XArray functionality
         >>>
@@ -708,7 +728,7 @@ def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = 
 
     Example THEMIS:
         >>> # Import necessary libraries
-        >>> import cdflib
+        >>> import cdflib.xarray
         >>> import xarray as xr
         >>> import os
         >>> import urllib.request
@@ -720,7 +740,7 @@ def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = 
         >>>     urllib.request.urlretrieve(url, fname)
 
         >>> # Load in and display the CDF file
-        >>> thg_data = cdflib.cdf_to_xarray(fname, to_unixtime=True, fillval_to_nan=True)
+        >>> thg_data = cdflib.xarray.cdf_to_xarray(fname, to_unixtime=True, fillval_to_nan=True)
 
     Processing Steps:
         1. For each variable in the CDF file
@@ -742,6 +762,10 @@ def cdf_to_xarray(filename: str, to_datetime: bool = False, to_unixtime: bool = 
         3. Gather all global scope attributes in the CDF file
         4. Create an XArray Dataset objects with the data variables, coordinate variables, and global attributes.
     """
+
+    if to_datetime and to_unixtime:
+        to_datetime = False
+
     # Convert the CDF file into a series of dicts, so we don't need to keep reading the file
     global_attributes, all_variable_attributes, all_variable_data, all_variable_properties = _convert_cdf_to_dicts(
         filename, to_datetime=to_datetime, to_unixtime=to_unixtime